+++ docs/changelog.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### UNRELEASED
+### 0.21.0 - 2020-04-12
 
  - Create feature to pass kwargs through middlewares and controller